pyproject.toml
[tool.pdm]
[tool.pdm.dev-dependencies]
dev = [
    "isort>=5.12.0",
    "black>=23.7.0",
]

[project]
name = "livrarias-anas"
version = "0.1.0"
description = ""
authors = [
    {name = "Anas 3info1", email = "aninha10.desouza@gmail.com"},
]
dependencies = [
    "django>=4.1.7",
    "djangorestframework>=3.14.0",
    "django-cors-headers>=4.2.0",
    "djangorestframework-simplejwt>=5.2.2",
    "setuptools>=68.0.0",
    "python-magic>=0.4.27",
    "Pillow>=10.0.0",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-pep517>=1.0"]
build-backend = "pdm.pep517.api"

[[tool.pdm.autoexport]]
filename = "requirements.txt"
without-hashes = "true"

